Refactor `model.jl` into several files and separate module (#306)
