+++ .github/workflows/Tests.yml
@@ -17,18 +17,18 @@ jobs:
       matrix:
         version:
           - '1'
-          - 'min'
-          - 'pre'
+          # - 'min'
+          # - 'pre'
         os:
           - ubuntu-latest
-          - windows-latest
+          # - windows-latest
         arch:
           - x64
-        include:
-          - version: 'min'
-            os: ubuntu-latest
-            arch: x64
-            coverage: true
+        # include:
+        #   - version: 'min'
+        #     os: ubuntu-latest
+        #     arch: x64
+        #     coverage: true
 
     steps:
       - uses: actions/checkout@v4
@@ -75,10 +75,10 @@ jobs:
         env:
           TEST_GROUP: "log_density"
 
-      - name: Running `source_gen` tests
-        uses: julia-actions/julia-runtest@v1
-        env:
-          TEST_GROUP: "source_gen"
+      # - name: Running `source_gen` tests
+      #   uses: julia-actions/julia-runtest@v1
+      #   env:
+      #     TEST_GROUP: "source_gen"
 
       # - name: Running `gibbs` tests
       #   uses: nick-fields/retry@v3
@@ -103,16 +103,16 @@ jobs:
       #   env:
       #     TEST_GROUP: "experimental"
       
-      - uses: julia-actions/julia-processcoverage@v1
-        if: matrix.coverage
+      # - uses: julia-actions/julia-processcoverage@v1
+      #   if: matrix.coverage
 
-      - uses: codecov/codecov-action@v4
-        if: matrix.coverage
-        with:
-          file: lcov.info
+      # - uses: codecov/codecov-action@v4
+      #   if: matrix.coverage
+      #   with:
+      #     file: lcov.info
       
-      - uses: coverallsapp/github-action@master
-        if: matrix.coverage
-        with:
-          github-token: ${{ secrets.GITHUB_TOKEN }}
-          path-to-lcov: lcov.info
+      # - uses: coverallsapp/github-action@master
+      #   if: matrix.coverage
+      #   with:
+      #     github-token: ${{ secrets.GITHUB_TOKEN }}
+      #     path-to-lcov: lcov.info
