+++ .github/workflows/TestsMacOS.yml
@@ -16,7 +16,10 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        version: ['1', 'min', 'pre']
+        version: 
+          - '1'
+          # - 'min'
+          # - 'pre'
         os: [macOS-latest]
         arch: [aarch64]
 
@@ -65,10 +68,10 @@ jobs:
         env:
           TEST_GROUP: "log_density"
 
-      - name: Running `source_gen` tests
-        uses: julia-actions/julia-runtest@v1
-        env:
-          TEST_GROUP: "source_gen"
+      # - name: Running `source_gen` tests
+      #   uses: julia-actions/julia-runtest@v1
+      #   env:
+      #     TEST_GROUP: "source_gen"
 
       # - name: Running `gibbs` tests
       #   uses: nick-fields/retry@v3
