+++ src/logdensityproblems.jl
@@ -1,8 +1,8 @@
-function _eval_logdensity(model, ::UseGeneratedLogDensityFunction, x)
+function _eval_logdensity(model, ::Model.UseGeneratedLogDensityFunction, x)
     return model.log_density_computation_function(model.evaluation_env, x)
 end
 
-function _eval_logdensity(model, ::UseGraph, x)
+function _eval_logdensity(model, ::Model.UseGraph, x)
     _, logp = JuliaBUGS.evaluate!!(model, x)
     return logp
 end
@@ -19,6 +19,6 @@ function LogDensityProblems.dimension(model::BUGSModel)
     end
 end
 
-function LogDensityProblems.capabilities(::AbstractBUGSModel)
+function LogDensityProblems.capabilities(::Model.AbstractBUGSModel)
     return LogDensityProblems.LogDensityOrder{0}()
 end
