+++ test/passes.jl
@@ -171,5 +171,5 @@ end
         model_def, data, (;)
     )
     model = compile(model_def, data, (;))
-    @test length(JuliaBUGS.parameters(model)) == 0
+    @test length(model.parameters) == 0
 end
